feat: integrate Incentive API with transaction processing
- Added RestTemplate configuration
- Implemented IncentiveService for external API call
- Updated TransactionService to include incentive logic
- Adjusted balance calculation (recipient gets incentive, sender unaffected)
- Stored incentive in TransactionRecord
- Added debug logs for user balance tracking

+++ src/main/java/com/jpmc/midascore/entity/UserRecord.java
@@ -40,8 +40,8 @@ public class UserRecord {
         return balance;
     }
 
-    public void setBalance(float balance) {
-        this.balance = balance;
+    public void setBalance(double balance) {
+        this.balance = (float) balance;
     }
 
     public UserRecord orElse(Object object) {
